Move BatchOperations to their respective packages
