+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/TestProdId.kt
@@ -2,7 +2,7 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.ical4android.impl
+package at.bitfire.synctools
 
 import net.fortuna.ical4j.model.property.ProdId
