+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/storage/jtx/TestJtxIcalObject.kt
@@ -2,7 +2,7 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.ical4android.impl
+package at.bitfire.synctools.storage.jtx
 
 import android.content.ContentValues
 import at.bitfire.ical4android.JtxCollection
