--- synctools/src/main/kotlin/at/bitfire/synctools/storage/ContactsBatchOperation.kt
@@ -1,8 +1,0 @@
-/*
- * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
- */
-
-package at.bitfire.synctools.storage
-
-@Deprecated("Moved to at.bitfire.synctools.storage.contacts")
-typealias ContactsBatchOperation = at.bitfire.synctools.storage.contacts.ContactsBatchOperation
