+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/jtx/JtxBatchOperation.kt
@@ -2,12 +2,13 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.synctools.storage
+package at.bitfire.synctools.storage.jtx
 
 import android.content.ContentProviderClient
+import at.bitfire.synctools.storage.BatchOperation
 
 /**
- * [BatchOperation] for jtx Board
+ * [at.bitfire.synctools.storage.BatchOperation] for jtx Board
  */
 class JtxBatchOperation(
     providerClient: ContentProviderClient
